Fix intermittent errors in byomachine_controller
To mitigate intermittent errors when Patch operation fails,
we need to return empty result and an error, so that the reconcile loop
will continue to get to the desired state.

The Ready field of ByoMachineStatus has to be optional. Because its
value is set by the Patch operation, and a failure might complain that a
required field is missing