Fix flaky controller unit tests
Both unit and integration tests used the same reconciler object. The
flakes arise when we try to read an object that was just created using
k8sManager.GetClient() which read from cache. In this case, k8sClient
has to be used. But in the case where Indexing is needed, we cannot use
k8sClient, because that reads from the API Server and not the cache

As a temporary fix, using separate reconciler objects for both unit and
integration tests