TKG-7055 The version of byoh provider should be v0.1.0 (#100)
* TKG-7055 The version of byoh provider should be v0.1.0

In current e2e case, it used byoh:v0.4.0 which is not proper. Let’s change it to v0.1.0 as we are a new provider and we will follow our own version series.

This is a bug of e2e case. The version of byoh is v0.1.0, the version of other component is v0.4.2, they have different minor value, should not share one metada.yaml.

Signed-off-by: Chen Hui <huchen@vmware.com>