Renamed the HAI cli to cli-dev.go in accordance with issue #374 and corrected codecov with the new file name. (#387)
I explored the option to move cli.go to other folder to be more explicit that it is an optional workflow but it uses private methods and classes from the installer package. If it is moved we need to cut functionalities from the cli or refactor all other files which in turn leads to other problems or conflicts. With this in mind I decided that it is more viable option to just rename it at this time.