Create kubeconfig in bootstrap flow
Added changes for creation of kubeconfig if the CSR is approved. This
commit also introduces refactoring to the existing implementation.
BootstrapAuthenticator controller is removed as it could be better
handled in the csr.go code itself. The flow will wait till the timeout
for certificate approval and if approved before the timeout will create
kubeconfig. Some motivation is taken from the kubelet bootstrapping code
and some client-go APIs are used as well.